arm: stm32_spi: Increase size of DMA channel variables to 16 bits
The STM32L4 requires more bits for saving the DMA channel.

Reviewed-on: https://gerrit.mot.com/908402
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>

+++ nuttx/arch/arm/src/stm32/stm32_spi.c
@@ -203,8 +203,8 @@ struct stm32_spidev_s
 #ifdef CONFIG_STM32_SPI_DMA
   volatile uint8_t rxresult;   /* Result of the RX DMA */
   volatile uint8_t txresult;   /* Result of the RX DMA */
-  uint8_t          rxch;       /* The RX DMA channel number */
-  uint8_t          txch;       /* The TX DMA channel number */
+  uint16_t         rxch;       /* The RX DMA channel number */
+  uint16_t         txch;       /* The TX DMA channel number */
   DMA_HANDLE       rxdma;      /* DMA channel handle for RX transfers */
   DMA_HANDLE       txdma;      /* DMA channel handle for TX transfers */
   sem_t            rxsem;      /* Wait for RX DMA to complete */
